fix number of ways

+++ number_of_ways.cpp
@@ -1,17 +1,17 @@
 #include <vector>
 #include <iostream>
 
-uint64_t num_ways(std::vector<uint64_t> test, uint64_t sum) {
+int64_t num_ways(std::vector<int64_t> test, int64_t sum) {
     //if the reminder of sum / 3 is != 0, return 0
     if(sum % 3)
         return 0;
 
     sum /= 3;
-    std::vector<int> aux(test.size());
-    int tmp = 0;
+    std::vector<int64_t> aux(test.size());
+    int64_t tmp = 0;
 
     //count the number of suffixes of test which sum to "sum"
-    for(int j = test.size() - 1; j >= 0; j--) {
+    for(int64_t j = test.size() - 1; j >= 0; j--) {
         tmp += test.at(j);
         if(tmp == sum) {
             aux.at(j) = 1;
@@ -19,16 +19,16 @@ uint64_t num_ways(std::vector<uint64_t> test, uint64_t sum) {
             aux.at(j) = 0;
     }
 
-    for(int j = test.size() - 2; j >= 0; j--) {
+    for(int64_t j = test.size() - 2; j >= 0; j--) {
         aux.at(j) += aux.at(j+1);
     }
 
-    uint64_t num_ways = 0;
+    int64_t num_ways = 0;
     tmp = 0;
 
     //compute the sum of prefixes of test
     //if a prefix sums to "sum", add aux[i+2] to the result
-    for(int j = 0; j < test.size() - 2; j++) {
+    for(int64_t j = 0; j < test.size() - 2; j++) {
         tmp += test.at(j);
         if(tmp == sum) {
             num_ways += aux.at(j+2);
@@ -39,12 +39,12 @@ uint64_t num_ways(std::vector<uint64_t> test, uint64_t sum) {
 }
 
 int main() {
-    int size;
-    std::vector<uint64_t> test;
+    int64_t size;
+    std::vector<int64_t> test;
     std::cin >> size;
     test.reserve(size);
-    int el;
-    uint64_t sum = 0;
+    int64_t el;
+    int64_t sum = 0;
 
     while(size--) {
         std::cin >> el;
